% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute MSE - Mean Squared Error}
\usage{
compute_mse(y_observed, y_predicted)
}
\arguments{
\item{y_observed}{[numeric] Observed values for the response variable}

\item{y_predicted}{[numeric] Predicted values for the response variable}
}
\value{
[numeric] The MSE produced by the linear model
}
\description{
Compute MSE - Mean Squared Error
}
