% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kfold_cv_lm_seq}
\alias{kfold_cv_lm_seq}
\title{K-fold cross-validation for lm - sequential version}
\usage{
kfold_cv_lm_seq(X, y, k = 5)
}
\arguments{
\item{X}{[numeric] Design matrix}

\item{y}{[numeric] Observed response vector}

\item{k}{[numeric] Number of folds}
}
\value{
[numeric] MSE provided by lm() R function through k-fold Cross-Validation
}
\description{
K-fold cross-validation for lm - sequential version
}
