% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{kfold_cv_parallel}
\alias{kfold_cv_parallel}
\title{kfold_cv_parallel}
\usage{
kfold_cv_parallel(
  X,
  y,
  k = 5,
  algorithm = "sd",
  tol = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{X}{[numeric] Design matrix}

\item{y}{[numeric]Response vector}

\item{k}{[numeric] Number of folds}

\item{algorithm}{[string] "sd" for steepest descent method; "gd" for gradient descent method}

\item{tol}{[numeric] Tolerance for stopping criteria}

\item{maxit}{[numeric] Maximum number of iterations for stopping criteria}

\item{stepsize}{[numeric] Stepsize}

\item{verbose}{[logical] verbose = TRUE prints some information on errors and number of iterations}
}
\value{
[numeric] Mean Squared Error
}
\description{
This function provides an MSE estimate of linear model predictions through a parallel k-fold cross validation method.
}
