% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{kfold_cv_seq}
\alias{kfold_cv_seq}
\title{kfold_cv_seq}
\usage{
kfold_cv_seq(
  X,
  y,
  k = 5,
  algorithm = "sd",
  tol = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{X}{[matrix] matrix of the covariates}

\item{y}{[vector] vector of the respose variable}

\item{k}{[numeric] the number of folds}

\item{algorithm}{[string] specifies "sd" for the steepest descent method or "gd" for the gradient descent method}

\item{tol}{[numeric] tolerance for the stoppage criteria of the algorithm}

\item{maxit}{[numeric] maximum number of iterations of the algorithm when convergence is not reached}

\item{stepsize}{[numeric]}

\item{verbose}{[logical]}
}
\value{
[numeric] the function returns the Mean Square Error of the model
}
\description{
This function provides a linear model estimate through a k-fold cross validation method.
}
