% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gradient}
\alias{gradient}
\title{Gradient of the loss function}
\usage{
gradient(beta, X, y)
}
\arguments{
\item{beta}{[numeric] Point where to evaluate the gradient}

\item{X}{[numeric] Design matrix}

\item{y}{[numeric] Observed response vector}
}
\value{
[numeric] Gradient of loss function in linear models
}
\description{
Gradient of the loss function
}
