% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gradient}
\alias{gradient}
\title{Gradient}
\usage{
gradient(beta, X, y)
}
\arguments{
\item{beta}{[numeric] Point where to evaluate the gradient}

\item{X}{[numeric] Design matrix}

\item{y}{[numeric] Data}
}
\value{
[numeric] The gradient of the function
}
\description{
Gradient
}
