% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{kfold_cv_seq}
\alias{kfold_cv_seq}
\title{K-fold cross-validation - sequential version}
\usage{
kfold_cv_seq(
  X,
  y,
  k = 5,
  algorithm = "sd",
  tol = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{X}{[matrix] Design matrix}

\item{y}{[vector] Observed response vector}

\item{k}{[numeric] Number of folds - default is 5}

\item{algorithm}{[string] "sd" for steepest descent method; "gd" for gradient descent method}

\item{tol}{[numeric] Tolerance for stopping criteria}

\item{maxit}{[numeric] Maximum number of iterations for stopping criteria}

\item{stepsize}{[numeric] Stepsize - not used for steepest descend method}

\item{verbose}{[logical]}
}
\value{
[numeric] The k-fold cross-validation MSE - Mean Squared Error
}
\description{
This function provides the MSE estimate of linear model predictions through the k-fold cross-validation method.
}
