% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{linear_gd_optim}
\usage{
linear_gd_optim(
  beta0,
  X,
  y,
  tol = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = T
)
}
\arguments{
\item{beta0}{[numeric] Initial values of coefficients}

\item{X}{[numeric] Matrix design}

\item{y}{[numeric] Response vector}

\item{tol}{[numeric] Tolerance to implement a stop rule for the algorithm}

\item{maxit}{[numeric] Maximum number of iterations}

\item{stepsize}{[numeric] Stepsize}

\item{verbose}{[logical] TRUE if you want display messages; FALSE otherwise}
}
\value{
[numeric] estimated coefficients through gradient algorithm
}
\description{
linear_gd_optim
}
