% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent.R
\name{linear_gd_optim}
\alias{linear_gd_optim}
\title{linear_gd_optim}
\usage{
linear_gd_optim(
  beta0,
  X,
  y,
  tol = 0.001,
  maxit = 1000,
  stepsize = 0.001,
  verbose = T
)
}
\arguments{
\item{beta0}{[numeric] Starting point of the iterative algorithm}

\item{X}{[numeric] Design matrix}

\item{y}{[numeric] Response vector}

\item{tol}{[numeric] Tolerance for stopping criteria}

\item{maxit}{[numeric] Maximum number of iterations for stopping criteria}

\item{stepsize}{[numeric] Stepsize}

\item{verbose}{[logical] verbose = TRUE prints some information on errors and number of iterations}
}
\value{
[numeric] Estimated minimum point via gradient algorithm
}
\description{
This function returns the minimum point via gradient algorithm
}
