% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steepest_descent.R
\name{linear_sd_optim}
\alias{linear_sd_optim}
\title{linear_sd_optim}
\usage{
linear_sd_optim(beta0, X, y, tol = 0.001, maxit = 1000, verbose = T)
}
\arguments{
\item{beta0}{[numeric] Starting point of the iterative algorithm}

\item{X}{[numeric] Design matrix}

\item{y}{[numeric] Response vector}

\item{tol}{[numeric] Tolerance for stopping criteria}

\item{maxit}{[numeric] Maximum number of iterations for stopping criteria}

\item{verbose}{[character] verbose = TRUE prints some information on errors and number of iterations}
}
\value{
[numeric] Estimated minimum point via steepest descend algorithm
}
\description{
This function returns the minimum point via steepest descend algorithm
}
