% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mymap}
\alias{mymap}
\title{Title}
\usage{
mymap(
  i_test,
  X_perm,
  y_perm,
  end_index_folds,
  algorithm,
  tol,
  maxit,
  stepsize,
  verbose
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
